chanegd by Abdulla

+++ apps/products/models.py
@@ -219,7 +219,9 @@ class ProductInventory(models.Model):
         decimal_places=2,
         verbose_name=_('price')
     )
-    sale_price = models.PositiveIntegerField(
+    sale_price = models.DecimalField(
+        max_digits=12,
+        decimal_places=2,
         verbose_name=_('sale_price')
     )
     installment_plan = models.CharField(max_length=250,verbose_name=_('installment_plan'))
